+++ src/index.js
@@ -5,7 +5,7 @@ export { default as Dropdown } from "./Dropdown";
 export { default as Button } from "./Button";
 export { default as Box } from "./Box";
 export { default as Menu } from "./Menu";
-export { default as SideSheet } from "./SideSheet";
+export { default as Sidesheet } from "./Sidesheet";
 export { default as Skeleton } from "./Skeleton";
 export * from "./Flags";
 export * from "./OperatingSystems";
